update after refactor to shapeshift
