--- demo-app/pod/.internal/setup/Y3VycmVudC1iYXNlLXVybA==$.json
@@ -1,1 +1,0 @@
-"http://localhost:3000/"
