--- demo-app/pod/.internal/setup/Y3VycmVudC1zZXJ2ZXItdmVyc2lvbg==$.json
@@ -1,1 +1,0 @@
-"5.1.2"
