--- demo-app/pod/.internal/setup/cm9vdEluaXRpYWxpemVk$.json
@@ -1,1 +1,0 @@
-true
